requirements.txt
numpy==1.20.1
pandas==1.2.3
py4j==0.10.9
pyspark==3.1.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
